+++ lib/src/bones_ui_utils.dart
@@ -1,7 +1,7 @@
-import 'dart:html';
-
 import 'package:collection/collection.dart';
 import 'package:dom_builder/dom_builder.dart';
+
+import 'bones_ui_web.dart';
 
 /// Returns a [StackTrace] ensuring that no error will be thrown.
 StackTrace stackTraceSafe() {
@@ -43,7 +43,7 @@ String? resolveToText(Object? o) {
   } else if (o is Iterable) {
     var l = o.whereNotNull().map(resolveToText).whereNotNull().toList();
     return l.isEmpty ? null : l.join();
-  } else if (o is Element) {
+  } else if (o is UIElement) {
     return o.text;
   } else if (o is DOMElement) {
     return o.text;